Major revamp of website
This is a major revamp of the HarlonTransit website, updating the content to reflect recent events such as the reopening of Metro Line M31 and the opening of Metro Line M5, and the suspension of the Commuter network.

New pages have been added for Metro  Line M31 and Metro Line M5 with details about stations, features and employment opportunities.

The global alerts have been updated to reflect the Commuter network suspension and to advertise employment opportunities on Metro Line M31.

The network map has been updated to a new version created by Janukiii_, whilst previous versions of the network map are now available to view and download publicly.

Finally, the site name has been updated to include the HarlonTransit slogan, and a "back to top" button has been added sitewide.

A changelog system powered by git-cliff has also been added to this repository.